Implements login
