+++ app/google-services.json
@@ -1,0 +1,39 @@
+{
+  "project_info": {
+    "project_number": "540770784782",
+    "project_id": "karo-65f67",
+    "storage_bucket": "karo-65f67.appspot.com"
+  },
+  "client": [
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:540770784782:android:68c045a5b3324abff88e5f",
+        "android_client_info": {
+          "package_name": "com.example.karo"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "540770784782-mh6rqad658rj0nr2ils1b495jpb7etoa.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyAOLM_pMykXAb5M5ck4WyxaM3LqZ-i8g7E"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "540770784782-mh6rqad658rj0nr2ils1b495jpb7etoa.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
+    }
+  ],
+  "configuration_version": "1"
+}
