+++ app/src/main/java/com/example/karo/LoginActivity.kt
@@ -1,0 +1,78 @@
+package com.example.karo
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.OutlinedTextField
+import androidx.compose.material.Text
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.unit.dp
+import com.example.karo.ui.theme.KaroTheme
+
+class LoginActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        fun login() {
+            startActivity(Intent(this, MainActivity::class.java))
+            finish()
+        }
+
+        setContent {
+            KaroTheme {
+                var email by remember { mutableStateOf(TextFieldValue("")) }
+                var password by remember { mutableStateOf(TextFieldValue("")) }
+
+                Column(
+                    modifier = Modifier.fillMaxSize(),
+                    verticalArrangement = Arrangement.Center,
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Image(
+                        painter = painterResource(id = R.drawable.logo),
+                        contentDescription = "Logo",
+                        contentScale = ContentScale.FillBounds,
+                        modifier = Modifier.size(100.dp)
+                    )
+
+                    Spacer(modifier = Modifier.height(30.dp))
+
+                    OutlinedTextField(
+                        value = email, onValueChange = { email = it },
+                        label = { Text("Email address") },
+                        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Email),
+                        placeholder = { Text("marie.ndathi@khalifa.nabz") },
+                    )
+
+                    Spacer(modifier = Modifier.height(10.dp))
+
+                    OutlinedTextField(
+                        value = password, onValueChange = { password = it },
+                        label = { Text("Password") },
+                        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
+                        placeholder = { Text("passwwaaddd") },
+                    )
+
+                    Spacer(modifier = Modifier.height(20.dp))
+
+                    ConfirmButton("Sign In") {
+                        login()
+                    }
+                }
+            }
+        }
+    }
+}
