+++ build.gradle
@@ -2,6 +2,9 @@ buildscript {
     ext {
         compose_ui_version = '1.1.1'
     }
+    dependencies {
+        classpath 'com.google.gms:google-services:4.3.14'
+    }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id 'com.android.application' version '7.3.1' apply false
